+++ linking-platform/link-provider/src/editor/__tests__/index.test.ts
@@ -1,7 +1,9 @@
 import { EditorCardProvider } from '..';
 import { ORSProvidersResponse } from '../types';
 
-const mockProvidersResponse: ORSProvidersResponse = {
+const getMockProvidersResponse = (
+  overridePolarisDefaultView: boolean = false,
+): ORSProvidersResponse => ({
   providers: [
     {
       key: 'google-object-provider',
@@ -9,7 +11,15 @@ const mockProvidersResponse: ORSProvidersResponse = {
         {
           source:
             '^https:\\/\\/docs.google.com\\/(?:spreadsheets|document|presentation)\\/d\\/[^\\\\]+\\/|^https:\\/\\/drive.google.com\\/file\\/d\\/[^\\\\]+\\/|^https:\\/\\/drive.google.com\\/open\\?id=[^&]+|^https:\\/\\/drive.google.com\\/drive\\/u\\/\\d+\\/folders\\/[^&\\?]+|^https:\\/\\/drive.google.com\\/drive\\/folders\\/[^&\\?]+',
-          flags: '',
+        },
+      ],
+    },
+    {
+      key: 'provider-with-default-view',
+      patterns: [
+        {
+          source: '^https:\\/\\/site-with-default-view.com\\/.*?/?$',
+          defaultView: 'embed',
         },
       ],
     },
@@ -19,17 +29,14 @@ const mockProvidersResponse: ORSProvidersResponse = {
         {
           source:
             '^https:\\/\\/.*?\\.jira-dev\\.com\\/browse\\/([a-zA-Z0-9]+-\\d+)#?.*?\\/?$',
-          flags: '',
         },
         {
           source:
             '^https:\\/\\/.*?\\.jira-dev\\.com\\/jira\\/software\\/(c\\/)?projects\\/([^\\/]+?)\\/boards\\/(\\d+)\\/roadmap\\/?',
-          flags: '',
         },
         {
           source:
             '^https:\\/\\/.*?\\.jira-dev\\.com\\/jira\\/core\\/projects\\/(?<resourceId>\\w+)\\/(timeline|calendar|list|board)\\/?',
-          flags: '',
         },
       ],
     },
@@ -39,7 +46,6 @@ const mockProvidersResponse: ORSProvidersResponse = {
         {
           source:
             '^https:\\/\\/.+?\\.slack\\.com\\/archives\\/[CG][A-Z0-9][^/]+\\/?$|^https:\\/\\/app\\.slack\\.com\\/client\\/T[A-Z0-9]+\\/[CG][A-Z0-9][^/]+\\/?$|^https:\\/\\/.+?\\.slack\\.com\\/archives\\/[CG][A-Z0-9][^/]+\\/p[0-9]+(\\?.*)?$',
-          flags: '',
         },
       ],
     },
@@ -49,12 +55,34 @@ const mockProvidersResponse: ORSProvidersResponse = {
         {
           source:
             '^https:\\/\\/.*?\\/jira\\/polaris\\/projects\\/[^\\/]+?\\/ideas\\/view\\/\\d+$|^https:\\/\\/.*?\\/secure\\/JiraProductDiscoveryAnonymous\\.jspa\\?hash=\\w+|^https:\\/\\/.*?\\/jira\\/polaris\\/share\\/\\w+',
-          flags: '',
+          ...(overridePolarisDefaultView ? { defaultView: 'block' } : {}),
         },
       ],
     },
   ],
-};
+});
+
+const expectedInlineAdf = (url: string) => ({
+  type: 'inlineCard',
+  attrs: {
+    url,
+  },
+});
+
+const expectedEmbedAdf = (url: string) => ({
+  type: 'embedCard',
+  attrs: {
+    url,
+    layout: 'wide',
+  },
+});
+
+const expectedBlockAdf = (url: string) => ({
+  type: 'blockCard',
+  attrs: {
+    url,
+  },
+});
 
 describe('providers > editor', () => {
   let mockFetch: jest.Mock;
@@ -74,292 +102,129 @@ describe('providers > editor', () => {
 
   it('returns inlineCard when calling /providers endpoint', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
       ok: true,
-    }));
+    });
     const url = 'https://drive.google.com/file/d/123/view?usp=sharing';
     const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'inlineCard',
-      attrs: {
-        url,
-      },
-    });
+    expect(adf).toEqual(expectedInlineAdf(url));
   });
 
   it('returns blockCard when calling /providers endpoint', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
       ok: true,
-    }));
+    });
     const url = 'https://drive.google.com/file/d/123/view?usp=sharing';
     const adf = await provider.resolve(url, 'block');
-    expect(adf).toEqual({
-      type: 'blockCard',
-      attrs: {
-        url,
-      },
-    });
+    expect(adf).toEqual(expectedBlockAdf(url));
   });
 
-  it('returns embedCard when roadmap embed inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://jdog.jira-dev.com/jira/software/projects/DL39857/boards/3186/roadmap';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
+  it.each<[string, string]>([
+    [
+      'Slack message',
+      'https://atlassian.slack.com/archives/C014W1DTRHS/p1614244582005100',
+    ],
+    [
+      'Slack message in thread',
+      'https://atlassian.slack.com/archives/C014W1DTRHS/p1614306173007200?thread_ts=1614244582.005100&cid=C014W1DTRHS',
+    ],
+  ])(
+    'returns inline when %s link inserted, calling /providers endpoint',
+    async (_, url) => {
+      const provider = new EditorCardProvider();
+      mockFetch.mockResolvedValueOnce({
+        json: async () => getMockProvidersResponse(),
+        ok: true,
+      });
+      const adf = await provider.resolve(url, 'inline');
+      expect(adf).toEqual(expectedInlineAdf(url));
+    },
+  );
 
-  it('returns embedCard when roadmap embed with query parameter inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://jdog.jira-dev.com/jira/software/projects/DL39857/boards/3186/roadmap?shared=&atlOrigin=eyJpIjoiYmFlNzRlMzAyYjAyNDlkZTgxZDc5ZTIzYmNlZmI5MjAiLCJwIjoiaiJ9';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when classic roadmap embed inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://jdog.jira-dev.com/jira/software/c/projects/DL39857/boards/3186/roadmap';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns inline when Slack message link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://atlassian.slack.com/archives/C014W1DTRHS/p1614244582005100';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'inlineCard',
-      attrs: {
-        url,
-      },
-    });
-  });
-
-  it('returns inline when Slack message in thread link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://atlassian.slack.com/archives/C014W1DTRHS/p1614306173007200?thread_ts=1614244582.005100&cid=C014W1DTRHS';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'inlineCard',
-      attrs: {
-        url,
-      },
-    });
-  });
-
-  it('returns embedCard when Polaris view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://polaris-v0.jira-dev.com/jira/polaris/projects/CS10/ideas/view/8981';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Polaris anonymous share view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://polaris-v0.jira-dev.com/jira/polaris/share/b2029c50914309acb37699615b1137da5';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Polaris anonymous share view fullscreen link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://polaris-v0.jira-dev.com/jira/polaris/share/89cb70599021ac29e227fc49c56782969?fullscreen=true';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Polaris anonymous resolved view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url =
-      'https://polaris-v0.jira-dev.com/secure/JiraProductDiscoveryAnonymous.jspa?hash=b2029c50914309acb37699615b1137da5';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Jira work management (JWM) timeline view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url = 'https://jdog.jira-dev.com/jira/core/projects/NPM5/timeline';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Jira work management (JWM) calendar view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url = 'https://jdog.jira-dev.com/jira/core/projects/NPM5/calendar';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Jira work management (JWM) list view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url = 'https://jdog.jira-dev.com/jira/core/projects/NPM5/list';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
-
-  it('returns embedCard when Jira work management (JWM) board view link inserted, calling /providers endpoint', async () => {
-    const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-      ok: true,
-    }));
-    const url = 'https://jdog.jira-dev.com/jira/core/projects/NPM5/board';
-    const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'embedCard',
-      attrs: {
-        url,
-        layout: 'wide',
-      },
-    });
-  });
+  it.each<[string, string]>([
+    [
+      'roadmap embed',
+      'https://jdog.jira-dev.com/jira/software/projects/DL39857/boards/3186/roadmap',
+    ],
+    [
+      'roadmap embed with query parameter',
+      'https://jdog.jira-dev.com/jira/software/projects/DL39857/boards/3186/roadmap?shared=&atlOrigin=eyJpIjoiYmFlNzRlMzAyYjAyNDlkZTgxZDc5ZTIzYmNlZmI5MjAiLCJwIjoiaiJ9',
+    ],
+    [
+      'classic roadmap embed',
+      'https://jdog.jira-dev.com/jira/software/c/projects/DL39857/boards/3186/roadmap',
+    ],
+    [
+      'Polaris view link',
+      'https://polaris-v0.jira-dev.com/jira/polaris/projects/CS10/ideas/view/8981',
+    ],
+    [
+      'Polaris anonymous share view',
+      'https://polaris-v0.jira-dev.com/jira/polaris/share/b2029c50914309acb37699615b1137da5',
+    ],
+    [
+      'Polaris anonymous share view fullscreen',
+      'https://polaris-v0.jira-dev.com/jira/polaris/share/89cb70599021ac29e227fc49c56782969?fullscreen=true',
+    ],
+    [
+      'Polaris anonymous resolved view',
+      'https://polaris-v0.jira-dev.com/secure/JiraProductDiscoveryAnonymous.jspa?hash=b2029c50914309acb37699615b1137da5',
+    ],
+    [
+      'Jira work management (JWM) timeline view',
+      'https://jdog.jira-dev.com/jira/core/projects/NPM5/timeline',
+    ],
+    [
+      'Jira work management (JWM) calendar view',
+      'https://jdog.jira-dev.com/jira/core/projects/NPM5/calendar',
+    ],
+    [
+      'Jira work management (JWM) list view',
+      'https://jdog.jira-dev.com/jira/core/projects/NPM5/list',
+    ],
+    [
+      'Jira work management (JWM) board view',
+      'https://jdog.jira-dev.com/jira/core/projects/NPM5/board',
+    ],
+  ])(
+    'returns embedCard when  %s link inserted, calling /providers endpoint',
+    async (_, url) => {
+      const provider = new EditorCardProvider();
+      mockFetch.mockResolvedValueOnce({
+        json: async () => getMockProvidersResponse(),
+        ok: true,
+      });
+      const adf = await provider.resolve(url, 'inline');
+      expect(adf).toEqual(expectedEmbedAdf(url));
+    },
+  );
 
   it('returns inlineCard when calling /providers endpoint, with fallback to /check', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
       ok: true,
-    }));
-    mockFetch.mockImplementationOnce(async () => ({
+    });
+    mockFetch.mockResolvedValueOnce({
       json: async () => ({ isSupported: true }),
       ok: true,
-    }));
+    });
     const url = 'https://drive.google.com/file/123';
     const adf = await provider.resolve(url, 'inline');
-    expect(adf).toEqual({
-      type: 'inlineCard',
-      attrs: {
-        url,
-      },
-    });
+    expect(adf).toEqual(expectedInlineAdf(url));
   });
 
   it('returns undefined when calling /providers endpoint, with fallback to /check, not supported', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
-    }));
-    mockFetch.mockImplementationOnce(async () => ({
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
+    });
+    mockFetch.mockResolvedValueOnce({
       json: async () => ({ isSupported: false }),
-    }));
+    });
     const url = 'https://drive.google.com/file/123';
     const promise = provider.resolve(url, 'inline');
     await expect(promise).rejects.toEqual(undefined);
@@ -367,16 +232,16 @@ describe('providers > editor', () => {
 
   it('returns undefined when calling /providers endpoint, with fallback to /check, both fail', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
+    mockFetch.mockResolvedValueOnce({
       json: async () => {
         throw Error();
       },
-    }));
-    mockFetch.mockImplementationOnce(async () => ({
+    });
+    mockFetch.mockResolvedValueOnce({
       json: async () => {
         throw Error();
       },
-    }));
+    });
     const url = 'https://drive.google.com/file/123';
     const promise = provider.resolve(url, 'inline');
     await expect(promise).rejects.toEqual(undefined);
@@ -384,15 +249,46 @@ describe('providers > editor', () => {
 
   it('calls /providers endpoint only once', async () => {
     const provider = new EditorCardProvider();
-    mockFetch.mockImplementationOnce(async () => ({
-      json: async () => mockProvidersResponse,
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
       ok: true,
-    }));
+    });
     await Promise.all([
       provider.resolve('https://drive.google.com/file/d/123/view', 'inline'),
       provider.resolve('https://drive.google.com/file/d/456/view', 'inline'),
       provider.resolve('https://drive.google.com/file/d/789/view', 'inline'),
     ]);
     expect(mockFetch).toHaveBeenCalledTimes(1);
+  });
+
+  it('should return EmbedCard when defaultView specifies it', async () => {
+    const provider = new EditorCardProvider();
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
+      ok: true,
+    });
+    const url = 'https://site-with-default-view.com/testing';
+    const adf = await provider.resolve(url, 'inline');
+    expect(adf).toEqual(expectedEmbedAdf(url));
+  });
+
+  it('should find pattern for a link', async () => {
+    const provider = new EditorCardProvider();
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
+      ok: true,
+    });
+    const url = 'https://site-with-default-view.com/testing';
+    expect(await provider.findPattern(url)).toBe(true);
+  });
+
+  it('should not find pattern for a link', async () => {
+    const provider = new EditorCardProvider();
+    mockFetch.mockResolvedValueOnce({
+      json: async () => getMockProvidersResponse(),
+      ok: true,
+    });
+    const url = 'https://site-without-pattern.com';
+    expect(await provider.findPattern(url)).toBe(false);
   });
 });