+++ linking-platform/link-provider/src/editor/index.ts
@@ -3,9 +3,15 @@
 import 'setimmediate';
 import DataLoader from 'dataloader';
 import { CardAdf, CardAppearance } from '@atlaskit/linking-common';
-import { CardProvider, ORSProvidersResponse, ORSCheckResponse } from './types';
+import {
+  CardProvider,
+  ORSCheckResponse,
+  ORSProvidersResponse,
+  ProviderPattern,
+} from './types';
 import { Transformer } from './transformer';
-import { getResolverUrl, getBaseUrl } from '../client/utils/environments';
+
+import { getBaseUrl, getResolverUrl } from '../client/utils/environments';
 import { EnvironmentsKeys } from '../client/types';
 import * as api from '../client/api';
 
@@ -14,10 +20,10 @@ const BATCH_WAIT_TIME = 50;
 export class EditorCardProvider implements CardProvider {
   private baseUrl: string;
   private resolverUrl: string;
-  private patterns?: string[];
+  private patterns?: ProviderPattern[];
   private requestHeaders: HeadersInit;
   private transformer: Transformer;
-  private patternLoader: DataLoader<string, string[] | undefined>;
+  private patternLoader: DataLoader<string, ProviderPattern[] | undefined>;
 
   constructor(envKey?: EnvironmentsKeys) {
     this.baseUrl = getBaseUrl(envKey);
@@ -33,7 +39,7 @@ export class EditorCardProvider implements CardProvider {
 
   private async batchPatterns(
     keys: ReadonlyArray<string>,
-  ): Promise<Array<string[] | undefined>> {
+  ): Promise<Array<ProviderPattern[] | undefined>> {
     // EDM-2205: Batch requests in the case that user paste multiple links at
     // once. This is so that only one /providers is being called.
     const patterns = await this.fetchPatterns();
@@ -59,7 +65,7 @@ export class EditorCardProvider implements CardProvider {
     }
   }
 
-  private async fetchPatterns(): Promise<string[] | undefined> {
+  private async fetchPatterns(): Promise<ProviderPattern[] | undefined> {
     try {
       const endpoint = `${this.resolverUrl}/providers`;
       const response = await api.request<ORSProvidersResponse>(
@@ -68,12 +74,13 @@ export class EditorCardProvider implements CardProvider {
         undefined,
         this.requestHeaders,
       );
-      return response.providers.reduce((allSources: string[], provider) => {
-        const providerSources = provider.patterns.map(
-          pattern => pattern.source,
-        );
-        return allSources.concat(providerSources);
-      }, []);
+
+      return response.providers.reduce(
+        (allSources: ProviderPattern[], provider) => {
+          return allSources.concat(provider.patterns);
+        },
+        [],
+      );
     } catch (err) {
       // eslint-disable-next-line
       console.error('failed to fetch /providers', err);
@@ -82,20 +89,27 @@ export class EditorCardProvider implements CardProvider {
   }
 
   async findPattern(url: string): Promise<boolean> {
+    return !!(await this.findPatternData(url));
+  }
+
+  private async findPatternData(
+    url: string,
+  ): Promise<ProviderPattern | undefined> {
     if (!this.patterns) {
       this.patterns = await this.patternLoader.load('providers');
     }
-    return this.patterns
-      ? this.patterns.some(pattern => url.match(pattern))
-      : false;
+    return this.patterns?.find(pattern => url.match(pattern.source));
   }
 
   async resolve(url: string, appearance: CardAppearance): Promise<CardAdf> {
     try {
-      let isSupported =
-        (await this.findPattern(url)) || (await this.check(url));
+      const matchedProviderPattern = await this.findPatternData(url);
+      let isSupported = !!matchedProviderPattern || (await this.check(url));
       if (isSupported) {
-        return this.transformer.toAdf(url, appearance);
+        return this.transformer.toAdf(
+          url,
+          matchedProviderPattern?.defaultView || appearance,
+        );
       }
     } catch (e) {
       // eslint-disable-next-line